refactor(rust cycle 17): split read_handlers 353->120 (+compose_helpers 236); send_endpoints 314->169 (+send_status 148)
