+++ src/bin/email_api_dir/mailbox/mod.rs
@@ -1,8 +1,10 @@
 // mailbox/mod.rs — re-exports uniquement (split Sprint 15)
 pub mod read_handlers;
+pub mod compose_helpers;
 pub mod send_pipeline;
 pub mod send_finalize;
 pub mod send_endpoints;
+pub mod send_status;
 pub mod send_queue_worker;
 pub mod single_handlers;
 pub mod drafts_handlers;
@@ -19,9 +21,11 @@ mod mime_attachments;
 pub(crate) use super::*;
 
 pub use read_handlers::*;
+pub use compose_helpers::*;
 pub use send_pipeline::*;
 pub use send_finalize::*;
 pub use send_endpoints::*;
+pub use send_status::*;
 pub use send_queue_worker::*;
 pub use single_handlers::*;
 pub use drafts_handlers::*;
