+++ src/bin/email_api_dir/mailbox/send_status.rs
@@ -1,0 +1,148 @@
+// Send status endpoint extracted from send_endpoints.rs (Sprint 17).
+#![allow(unused_imports)]
+use super::*;
+
+pub(crate) async fn api_send_status(
+    path: web::Path<String>,
+    req: actix_web::HttpRequest,
+    logic: web::Data<Arc<Logic>>,
+    mongo: web::Data<Arc<mongodb::Client>>,
+) -> impl Responder {
+    let user_id = resolve_user_id(&req);
+    let email_id = path.into_inner();
+
+    let email = match logic.fetch_email(&user_id, &email_id).await {
+        Ok(Some(email)) => email,
+        Ok(None) => {
+            return HttpResponse::NotFound().json(serde_json::json!({
+                "message": "Email not found"
+            }))
+        }
+        Err(e) => {
+            eprintln!("api_send_status fetch_email error: {}", e);
+            return HttpResponse::InternalServerError().json(serde_json::json!({
+                "message": "Failed to fetch email"
+            }));
+        }
+    };
+
+    let message_id_header = email
+        .headers
+        .iter()
+        .find(|(k, _)| k.eq_ignore_ascii_case("message-id"))
+        .map(|(_, v)| v.clone())
+        .unwrap_or_default();
+
+    let message_id = normalize_message_id(&message_id_header);
+    if message_id.is_empty() {
+        return HttpResponse::Ok().json(serde_json::json!({
+            "id": email.id,
+            "message": "No Message-ID found",
+            "monitoring": {
+                "traceable": false
+            }
+        }));
+    }
+
+    let db = env::var("MONGODB_DATABASE").unwrap_or_else(|_| "mailserver".to_string());
+    let coll = mongo
+        .database(&db)
+        .collection::<bson::Document>("smtp_events");
+
+    let events = match coll
+        .find(doc! { "message_id": &message_id })
+        .sort(doc! { "ts": -1 })
+        .limit(200)
+        .await
+    {
+        Ok(cursor) => cursor
+            .try_collect::<Vec<_>>()
+            .await
+            .unwrap_or_default()
+            .into_iter()
+            .filter_map(|d| bson::from_document::<monitoring::SmtpEvent>(d).ok())
+            .collect::<Vec<_>>(),
+        Err(e) => {
+            eprintln!("api_send_status query smtp_events error: {}", e);
+            vec![]
+        }
+    };
+
+    let accepted_by_remote_mx = events.iter().any(|e| {
+        matches!(e.event_type, monitoring::SmtpEventType::Delivered)
+            && matches!(e.status, monitoring::SmtpStatus::Delivered)
+            && !is_internal_delivery_hop(
+                e.mx_host.as_deref(),
+                e.remote_ip.as_deref(),
+                e.remote_port,
+                e.company.as_deref(),
+            )
+    });
+    let bounced_or_failed = events.iter().any(|e| {
+        matches!(
+            e.status,
+            monitoring::SmtpStatus::Bounced | monitoring::SmtpStatus::Failed
+        )
+    });
+    let saw_internal_handoff = events.iter().any(|e| {
+        is_internal_delivery_hop(
+            e.mx_host.as_deref(),
+            e.remote_ip.as_deref(),
+            e.remote_port,
+            e.company.as_deref(),
+        )
+    });
+    let handoff_only = !accepted_by_remote_mx && !bounced_or_failed && saw_internal_handoff;
+
+    let latest = events
+        .iter()
+        .find(|e| {
+            matches!(
+                e.status,
+                monitoring::SmtpStatus::Delivered
+                    | monitoring::SmtpStatus::Bounced
+                    | monitoring::SmtpStatus::Failed
+                    | monitoring::SmtpStatus::Deferred
+            )
+        })
+        .or_else(|| events.first());
+    let delivery_state = if accepted_by_remote_mx {
+        "sent"
+    } else if bounced_or_failed {
+        "failed"
+    } else if handoff_only {
+        "queued"
+    } else {
+        "sending"
+    };
+
+    HttpResponse::Ok().json(serde_json::json!({
+        "id": email.id,
+        "messageId": message_id,
+        "deliveryState": delivery_state,
+        "from": email.from,
+        "to": email.to,
+        "subject": email.subject,
+        "monitoring": {
+            "traceable": true,
+            "events": events.len(),
+            "acceptedByRemoteMx": accepted_by_remote_mx,
+            "bouncedOrFailed": bounced_or_failed,
+            "handoffOnly": handoff_only,
+            "latestEventType": latest.map(|e| format!("{:?}", e.event_type)),
+            "latestStatus": latest.map(|e| format!("{:?}", e.status)),
+            "latestSmtpCode": latest.and_then(|e| e.smtp_code),
+            "latestSmtpReply": latest.and_then(|e| e.smtp_reply.clone()),
+            "traceEndpoint": format!("/api/monitoring/messages/{}/trace", message_id),
+            "note": if accepted_by_remote_mx {
+                "Remote MX accepted the message (strong delivery signal)."
+            } else if handoff_only {
+                "Message handed off to DKIM service; remote mailbox receipt is not independently verified yet."
+            } else if bounced_or_failed {
+                "SMTP monitoring reports bounce/failure events."
+            } else {
+                "No conclusive delivery signal yet."
+            }
+        }
+    }))
+}
